+++ documents-webapp/src/main/webapp/vue-app/documents-user-setting-common/js/ApiKeyService.js
@@ -15,17 +15,6 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-export async function sendOtpCode(optMethod) {
-  const resp = await fetch(`/social/rest/apiKey?method=${optMethod}`, {
-    method: 'GET',
-    credentials: 'include'
-  });
-  if (!resp?.ok) {
-    const msg = resp ? await resp.text() : 'Unkown error';
-    throw new Error(`Server Error: ${msg}`);
-  }
-}
-
 export async function getPassword(optMethod, otpCode, renew) {
   const formData = new FormData();
   formData.append('method', optMethod);
